mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.57.1",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "mcp>=1.11.0",
    "python-dotenv>=1.1.1",
    "youtube-transcript-api>=1.1.1",
]
